requirements.txt
flask>=2.3.0
Pillow>=10.0.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
roboflow>=1.0.0  # For downloading vehicle detection datasets
